requirements.txt
blue==0.7.0
isort==5.10.1
mypy==0.931
pip-chill==1.0.1
pygithub==1.55
pytest-cov==3.0.0
requests-mock==1.9.3
types-requests==2.27.7
